requirements.txt
# Core Streamlit app
streamlit

# Data processing
pandas
sqlalchemy==2.0.45

# Visualization
plotly

# LangChain ecosystem (for LLM agent)
langchain==1.2.0
langchain-community==0.4.1
langchain-google-genai==4.1.2
langchain-experimental==0.4.1

# Monitoring
langsmith

# Google Gemini API (dependency of langchain-google-genai)

# Environment variable management
python-dotenv==1.2.1

# Optional: For PostgreSQL support
psycopg2

# Optional: For file/folder path handling
pathlib

time